AIchatbot/requirements.txt
nltk==3.7
Pillow==8.3.1
torch==2.0.1
numpy==1.21.5